backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
email-validator==2.1.0.post1
pydantic[email]==2.6.3
bcrypt==3.2.2
httpx==0.27.0
pytest==8.0.2
pytest-asyncio==0.23.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
email-validator==2.1.0.post1
pydantic[email]==2.6.3
bcrypt==3.2.2
httpx==0.27.0
